+++ src/main.rs
@@ -1,6 +1,5 @@
 use preexplorer::prelude::*;
 
 fn main() {
-	println!("{:?}", (vec![1, 2, 3]).preexplore());
-
+    println!("{:?}", (vec![1, 2, 3]).preexplore());
 }